LayoutTests/fast/dom/dom-parse-close-form-expected.txt

This tests that DOMParser closes a form element when it sees its close tag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new DOMParser()).parseFromString("<form></form>text", "text/html").body.innerHTML is "<form></form>text"
PASS successfullyParsed is true

TEST COMPLETE
